requirements.txt
Django==1.7
gunicorn==19.1
django-environ==0.3
